third_party/blink/web_tests/platform/mac-mac14-arm64/external/wpt/fetch/fetch-later/new-window.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] A cross-origin window[target=''][features=''] can trigger fetchLater.
  assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
[FAIL] A cross-origin window[target=''][features='popup'] can trigger fetchLater.
  assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
[FAIL] A cross-origin window[target='_blank'][features=''] can trigger fetchLater.
  assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
[FAIL] A cross-origin window[target='_blank'][features='popup'] can trigger fetchLater.
  assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
Harness: the test ran to completion.
